+++ ebpub/ebpub/neighbornews/models.py
@@ -18,7 +18,6 @@
 
 from django.contrib.gis.db import models
 from ebpub.accounts.models import User
-from ebpub.db.models import Lookup
 from ebpub.db.models import NewsItem
 
 class NewsItemCreator(models.Model):
@@ -41,29 +40,19 @@ class FeaturedLookupManager(models.Manager):
     def get_by_natural_key(self, slug):
         return self.get(lookup__slug=slug)
 
-
-class FeaturedLookup(models.Model):
-    """
-    This allows admins to designate some lookup values as special, for
-    eg. use in extra-prominent links, navigation categories, etc.
-    """
-
-    lookup = models.ForeignKey(Lookup, help_text='Which Lookup value to feature', unique=True)
-
-    @property
-    def name(self):
-        return self.lookup.name
-
-    @property
-    def code(self):
-        return self.lookup.code
-
-    @property
-    def slug(self):
-        return self.lookup.slug
-
-    objects = FeaturedLookupManager()
-
-    def natural_key(self):
-        return (self.lookup.slug,)
-
+    def featured_lookups_for(self, newsitem, sf_name):
+        """
+        Return a list of the Lookups that are featured and that the
+        newsitem has.
+        """
+        # This uses several queries, not very efficient...
+        from ebpub.db.models import SchemaField
+        sf = SchemaField.objects.get(schema__id=newsitem.schema_id, name=sf_name)
+        # Yet another manual decode of the comma-separated value,
+        # refs #265
+        if sf.is_many_to_many_lookup():
+            ni_lookup_ids = [int(i) for i in newsitem.attributes[sf_name].split(',')]
+        else:
+            ni_lookup_ids = [newsitem.attributes[sf_name]]
+        featured = self.filter(lookup__id__in=ni_lookup_ids).select_related()
+        return [obj.lookup for obj in featured]