+++ docs/changes/release_notes.rst
@@ -51,8 +51,13 @@ New Features in 1.2
 * Allow overriding the template for schema_filter view on a per-schema
   basis, by creating a template named db/schema_filter/<schema>.html
 
-* Added ``neighbornews.FeaturedLookup`` model that allows admins to
-  flag some Lookup values as "special", for use in eg. navigation
+* Added ``featured`` flag on ``ebpub.db.Lookup`` model, allowing admins
+  to designate some Lookup values as "special", for use in
+  eg. navigation.
+
+* Added ``Lookup.objects.get_featured_lookup_for_item(newsitem, attribute_key)``
+  method to find out which "featured" Lookups a newsitem has for a
+  given attribute.
 
 * Added ``ebpub.moderation`` app that allows users to flag NewsItems
   as spam or inappropriate, and an admin UI for it;